climate-lakehouse/requirements.txt
requests==2.32.0
python-dotenv==1.1.1
kafka-python==2.1.2
boto3
duckdb
deltalake
pyarrow